fix(Snap): add decimals to PM

+++ apps/snap/serializers/price_monitor_serializers.py
@@ -61,10 +61,10 @@ class UpdatePriceMonitorSnapSerializer(PriceMonitorSnapSerializer):
 
 class OverviewPriceMonitorSnapReport(serializers.Serializer):
     sku = serializers.CharField(source='sku_name')
-    min = serializers.IntegerField(source='min_value')
-    max = serializers.IntegerField(source='max_value')
-    mode = serializers.IntegerField(source='mode_value')
-    mean = serializers.IntegerField(source='mean_value')
+    min = serializers.DecimalField(max_digits=10, decimal_places=1, source='min_value')
+    max = serializers.DecimalField(max_digits=10, decimal_places=1, source='max_value')
+    mode = serializers.DecimalField(max_digits=10, decimal_places=1, source='mode_value')
+    mean = serializers.DecimalField(max_digits=10, decimal_places=1, source='mean_value')
 
 
 class MonthPriceMonitorSnapReportSerializer(serializers.Serializer):
